Move loadRecoveryData function to utils to enable usage in tests
